pyproject.toml
[project]
name = "arsidet"
version = "1.0.0"
description = "Real-time Arabic Sign Language Detection with AI-powered sentence interpretation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.7",
    "opencv-python>=4.12.0.88",
    "openai>=1.0.0",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scipy>=1.16.2",
    "seaborn>=0.13.2",
    "streamlit>=1.50.0",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "tqdm>=4.67.1",
    "ultralytics>=8.0.0",
]
